* Rename class * Migrate to JUnit5 * Format
